+++ frontend/.storybook/preview.ts
@@ -6,8 +6,8 @@ const preview: Preview = {
   parameters: {
     layout: 'centered',
     backgrounds: {
-      default: 'aircast-dark',
-      values: [{ name: 'aircast-dark', value: '#090d16' }],
+      default: 'adcast-dark',
+      values: [{ name: 'adcast-dark', value: '#090d16' }],
     },
     controls: { matchers: { color: /(background|color)$/i, date: /Date$/i } },
   },